requirements.txt
asgiref==3.8.1
chardet==5.2.0
crispy-tailwind==1.0.3
dj-database-url==2.3.0
Django==5.2
django-crispy-forms==2.4
django-environ==0.12.0
mysqlclient==2.2.7
pillow==11.2.1
psycopg2-binary==2.9.10
pytz==2025.2
reportlab==4.4.0
sqlparse==0.5.3
typing_extensions==4.13.2
